•	Security Upgrade for requests - upgraded to token excange with OSV server •	Recordings are uploaded via stream instead of opening the file in memory. This helps in two big ways:           o	Decreases the memory consumption while uploading           o	Progress UI is much more accurate and smooth. •	On/Off setting for displaying the map while recording •	Removed strong dependency to Google Play Services. It’s now optional. •	Display and SD card icon for tracks recorded on removable storage •	Fixed various reported bugs.

+++ app/src/main/java/com/telenav/osv/listener/UploadProgressListener.java
@@ -20,7 +20,7 @@ public interface UploadProgressListener extends ConnectionClassManager.Connectio
 
     void onUploadResumed();
 
-    void onUploadCancelled(int total, int remaining);
+    void onUploadCancelled(long total, long remaining);
 
     void onUploadFinished(int successful, int unsuccessful);
